tools/internal/parser: extract metadata enrichment from main parser
